LPD-42068 Moving to the right packages
