+++ modules/apps/headless/headless-batch-engine/headless-batch-engine-test/src/testIntegration/java/com/liferay/headless/batch/engine/entity/TestEntity.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.headless.batch.engine.resource.v1_0.test;
+package com.liferay.headless.batch.engine.entity;
 
 import java.io.Serializable;
