+++ modules/apps/headless/headless-batch-engine/headless-batch-engine-test/src/testIntegration/java/com/liferay/headless/batch/engine/vulcan/problem/TestEntityProblemMapper.java
@@ -3,8 +3,9 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.headless.batch.engine.resource.v1_0.test;
+package com.liferay.headless.batch.engine.vulcan.problem;
 
+import com.liferay.headless.batch.engine.exception.TestEntityException;
 import com.liferay.portal.vulcan.problem.Problem;
 import com.liferay.portal.vulcan.problem.ProblemMapper;
